+++ src/main/java/mg/fita/kinkinasaifu/repositories/OtherCrudOperations.java
@@ -16,10 +16,10 @@ public class OtherCrudOperations {
       "SELECT value FROM \"currency_value\" WHERE source_currency_id = ? "
           + "AND target_currency_id = ? ORDER BY effective_date LIMIT 1";
 
-  public Currency findById(int id) throws SQLException {
+  public Currency findById(long id) throws SQLException {
     Currency currency = null;
     try (PreparedStatement statement = connection.prepareStatement(FIND_CURRENCY_BY_ID_QUERY)) {
-      statement.setInt(1, id);
+      statement.setLong(1, id);
       try (ResultSet resultSet = statement.executeQuery()) {
         if (resultSet.next()) {
           currency =
@@ -33,12 +33,12 @@ public class OtherCrudOperations {
     return currency;
   }
 
-  public double getCurrencyValue(int sourceCurrencyId, int targetCurrencyId) throws SQLException {
+  public double getCurrencyValue(long sourceCurrencyId, long targetCurrencyId) throws SQLException {
     double currencyValue = 0.0;
     try (PreparedStatement statement =
         connection.prepareStatement(GET_MOST_RECENT_CURRENCY_VALUE)) {
-      statement.setInt(1, sourceCurrencyId);
-      statement.setInt(2, targetCurrencyId);
+      statement.setLong(1, sourceCurrencyId);
+      statement.setLong(2, targetCurrencyId);
       try (ResultSet resultSet = statement.executeQuery()) {
         if (resultSet.next()) {
           currencyValue = resultSet.getDouble(ColumnLabel.CurrencyValueTable.VALUE);
